backend/requirements.txt
flask
python-dotenv
azure-ai-formrecognizer
openai
werkzeug
flask_cors
pdf2image
PyMuPDF